whoops started gcing

+++ run.sh
@@ -1,3 +1,5 @@
-rm ./a.out
+if [ -f a.out ]; then
+  rm ./a.out
+fi
 gcc *.c
 ./a.out $1
